refactor import to sqlite logic
